Fix to munkicommon.getAvailableDiskSpace() so long int doesn't get written to ManagedInstallsReport.plist, which results in an invalid plist as far as MunkiReport is concerned.
git-svn-id: http://munki.googlecode.com/svn/trunk@1009 a4e17f2e-e282-11dd-95e1-755cbddbdd66

+++ code/client/munkilib/munkicommon.py
@@ -1597,7 +1597,7 @@ def getAvailableDiskSpace(volumepath='/'):
             'Error getting disk space in %s: %s', volumepath, str(e))
         return 0
 
-    return st.f_frsize * st.f_bavail / 1024 # f_bavail matches df(1) output
+    return int(st.f_frsize * st.f_bavail / 1024) # f_bavail matches df(1) output
 
 
 def cleanUpTmpDir():
